C: handler for abort (#135)
* C: handler for abort

CEF will kill the thread if the GPU process cannot be created. This
brings down Suwayomi with it, so instead just kill the thread and inform
Java about the exception.

* Two-stage build to download an build catch_abort

Building JNI extensions requires JDK, but we don't need it at runtime,
so create a build-step based on JDK, then copy the built binary over

* Add catch handler to github action