Improve image size and cachability (#165)
* Merge basic install layers

* Move `TACHIDESK_RELEASE_DOWNLOAD_URL` to where it's needed

This improves cachability, because the `ARG` taints all layers following
it. Since we change this arg for every release, the whole container had
to be rebuilt and redownloaded, when only the layer with the actual jar
file needs it.

* Move `ARG`s to the latest layer we can

`ARG`s taint every layer following, so move the `LABEL` to the last
layer, and the JAR download to the second-to-last, allowing for caching
of every layer except those two.

* Merge user creation layers and move `chmod`

The `chmod` needs to be after all copy and download if we want to avoid
permission problems (previous commit was not functional at runtime due
to insufficient permissions). This way, the files will be owned by the
correct user, and readable by all, while not creating unnecessary
layers.